requirements/test.txt
-r base.txt

coverage==3.7.1
django-nose==1.2
nose==1.3.0
